frontend/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-MySQLdb==1.0.1
python-dotenv==0.21.1
pyjwt==2.6.0
requests==2.28.2
Werkzeug==2.2.2
mysqlclient==2.2.4
argon2-cffi>=21.3.0
cryptography==46.0.4
